% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_selected.R
\docType{data}
\name{A_selected}
\alias{A_selected}
\title{Adjacency Matrix for Orthology Graph}
\format{
A matrix with 155 rows and 155 columns
}
\source{
It has been generated using the function orthG::mapTrees()
and the reconciliation output file 'selected'.
Verbigracia: orthG::mapTrees('./inst/extdata/selected')
The reconciliation was carried out using RANGER-DTL with parameters D = 1, T = 10 and L = 1.
}
\usage{
A_selected
}
\description{
155 x 155 square matrix (155 GS proteins from 45 seed plant species)
}
\keyword{datasets}
