% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AngGym.R
\docType{data}
\name{AngGym}
\alias{AngGym}
\title{Angiosperms Gymnosperms}
\format{
A dataframe with 155 rows (GS proteins) and 10 columns:
\describe{
\item{n}{Reference number}
\item{phylo_id}{Unique identification label of the protein/gen}
\item{species}{Species}
\item{taxon}{Acrogymnospermae or Angiospermae}
\item{class}{Angiosperms: Amborellopsida, Liliopsida, Magnoliopsida; Gymnosperms: Ginkgoopsida, Cycadopsida, Gnetopsida, Pinopsida}
\item{dna}{CDS sequence}
\item{prot}{Protein sequence}
\item{short}{Unique three letter identification of the species}
\item{gsLineage}{Either GS2, GS1a or GS1b}
\item{plant_group}{Primitive angiosperms, Modern angiosperms, Ginkgo-Cycadales, Gnetales, Pinacea, Conifer II}
}
}
\source{
It has been manually curated by the authors
}
\usage{
AngGym
}
\description{
Angiosperms Gymnosperms
}
\keyword{datasets}
