% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseRec.R
\name{coltips}
\alias{coltips}
\title{Colouring Tree Tips}
\usage{
coltips(phy)
}
\arguments{
\item{phy}{tree as a phylo object}
}
\value{
a color vector as long as the number of tips
}
\description{
Make a color vector for colouring tree tips
}
\details{
Each tip is given a color according to the nature of the isoform: green (GS2), blue (GS1a), brown (GS1b Gym), salmon (GS1b Ang), purple (other).
}
\examples{
coltips(ape::read.tree(text = "((Bdi, Sly), (Pp, Ap));"))
}
