% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth.R
\name{getseqGS}
\alias{getseqGS}
\title{Get the GS Sequence}
\usage{
getseqGS(phylo_id, molecule = "Prot")
}
\arguments{
\item{phylo_id}{the unique sequence identifier}

\item{molecule}{either "Prot" or "CDS"}
}
\value{
The requested sequence as a character string.
}
\description{
Provides the requested GS sequence
}
\details{
The identifier should be one of the 'phylo_id' from data(agf).
}
\examples{
getseqGS("Pp_GS1b_2")
}
