% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth.R
\name{orthG}
\alias{orthG}
\title{Infer GS OrthoGroups Within a Set of Species}
\usage{
orthG(set = "all")
}
\arguments{
\item{set}{set of species of interest provided as a character vector either with the binomial or short code of the species (see data(sdf)).}
}
\value{
A list with two elements. The first one is the adjacency matrix (1 for orthologous, 0 for paralogous). The second element is an orthogroup graph.
}
\description{
Infers GS orthogroups using tree reconciliation
}
\details{
When set = "all", all the species in the database will be included.
}
\examples{
orthG(set = c("Pp", "Psy", "Psm", "Ap"))
}
