% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth.R
\name{orthP}
\alias{orthP}
\title{Search Orthologous of a Given Protein}
\usage{
orthP(phylo_id, set = "all")
}
\arguments{
\item{phylo_id}{phylo_id of the query protein}

\item{set}{set of species of interest provided as a character vector, either with the binomial or short code of the species (see details).}
}
\value{
A list with thee elements: 1. subtree of the relevant proteins; 2. vector color; 3. phylo_ids of the orthologous found.
}
\description{
Searchs orthologous of a given protein within a set of selected species
}
\details{
When set = "all", the search will be carry out against all the species in the database.
}
\examples{
orthP(phylo_id = "Pp_GS1a", set = c("Pp", "Psy", "Psm", "Ap"))
}
