% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\docType{data}
\name{sdf}
\alias{sdf}
\title{Seed Plants and Ferns GS}
\format{
A dataframe with 196 rows (GS proteins) and 7 columns:
\describe{
\item{n}{Reference number}
\item{Sec.Name_}{Unique identification label of the protein}
\item{species}{Species}
\item{taxon}{Acrogymnospermae, Angiospermae  or Polypodiopsida}
\item{short}{Unique three letter identification of the species}
\item{gs}{Either GS2, GS1a, GS1b_Gym or GS1b_Ang. Here the ferns proteins have been forced to be either GS1a or GS2}
\item{tax_group}{Taxonomic group}
}
}
\source{
It has been handly curated by the authors
}
\usage{
sdf
}
\description{
155 GS proteins from 25 seed plants species and 41 GS proteins from 11 fern species
}
\keyword{datasets}
