% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected_tr.R
\docType{data}
\name{selected_tr}
\alias{selected_tr}
\title{Ultrametric Rooted Seed Plants Tree}
\format{
An phylo object
}
\source{
It has been manually curated by the authors
}
\usage{
selected_tr
}
\description{
155 GS proteins from 45 seed plants species
Rooted using MAD (Minimal Ancestor Deviation)
}
\keyword{datasets}
