% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orth.R
\name{subsetGS}
\alias{subsetGS}
\title{GS Proteins Report}
\usage{
subsetGS(sp)
}
\arguments{
\item{sp}{set of species of interest (either binomial or short code name)}
}
\value{
A dataframe with the information for the requested species.
}
\description{
Assembles a report regarding the GS proteins found in the indicated subset of species
}
\details{
This function returns the protein and DNA sequences of the different isoforms found in each species, along with other relevant data.
}
\examples{
subsetGS(c("Pinus pinaster", "Ath"))
}
